ChatPal/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord.py-self>=2.0.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.0",
    "websockets>=12.0",
    "colorama>=0.4.6",
    "flask>=3.0.0"
]
